robotics/analog_sensor: use strong typing
Turn 'int' type into 'uint16_t' type as it is always unsigned values.
As sum variable can overflow uint16_t before the average is done, it is
declared as uint32_t.

+++ robotics/analog_sensor.c
@@ -14,9 +14,9 @@ void *task_analog_sensors(void *arg)
 	{
 		uint8_t j, i;
 #if defined(AVERAGING)
-		int sum = 0;
+		uint32_t sum = 0;
 #endif
-		int value = 0;
+		uint16_t value = 0;
 
 		/* current index acquisition */
 		i = as->sensor_index;
@@ -62,7 +62,7 @@ void analog_sensor_setup(analog_sensors_t *as)
 	}
 }
 
-static dist_cm_t analog_sensor_adc2cm(int adc,
+static dist_cm_t analog_sensor_adc2cm(uint16_t adc,
 				    float coeff_volts, float const_volts,
 				    float const_dist, uint8_t dist_max)
 {
@@ -83,7 +83,7 @@ dist_cm_t analog_sensor_check_obstacle(analog_sensors_t *as,
 	dist_cm_t dist = AS_DIST_MAX;
 
 	if (id < as->sensors_nb) {
-		int raw;
+		uint8_t raw;
 
 		raw = as->sensors[id].raw_values[ANALOG_SENSOR_NB_SAMPLES-1];
 		dist = analog_sensor_adc2cm(raw,
